moving UpdateableCountryText
